+++ src/bin/xpg.rs
@@ -32,6 +32,10 @@ Notes
         * `keychain`
         * `codename`
         * `codename-series`
+
+2. By default, extended words are only available in exclusively extended word
+   kinds; however, with `-e`, they are merged with regular word kinds; use with
+   the `-L` option to see its effect.
 "
 )]
 struct Cli {
@@ -59,17 +63,24 @@ struct Cli {
     #[arg(short = 'C', long, value_name = "PATH")]
     config: Option<PathBuf>,
 
-    /// Print default configuration
+    /// Merge extended words
     #[arg(short, long)]
-    default_config: bool,
+    extended: bool,
+
+    /// List words in `{sub}`(s)
+    #[arg(short = 'L', long)]
+    list: bool,
+
+    /// Print configuration
+    #[arg(short, long, conflicts_with = "list")]
+    dump_config: bool,
 
     /// Print readme
     #[arg(short, long)]
     readme: bool,
 
-    /// Pattern (see note 1 below)
-    #[arg(default_value = "wwww")]
-    pattern: String,
+    /// Pattern(s) (see note 1 below) [default: `wwww`, `-L`: `{a}`]
+    patterns: Vec<String>,
 }
 
 /**
@@ -85,28 +96,45 @@ fn main() -> Result<()> {
         std::process::exit(0);
     }
 
-    if cli.default_config {
-        print!("{}", xpg::CONFIG.dump()?);
-        std::process::exit(0);
-    }
-
+    // Process the configuration
     let cfg = if let Some(config) = cli.config {
         // Load a custom configuration via -C
-        xpg::Config::from_path(&config)?
+        xpg::Config::from_path(&config, cli.extended)?
     } else if let Some(proj_dirs) = directories::ProjectDirs::from("com", "qtfkwk", "xpg") {
         let config_dir = proj_dirs.config_dir();
         let user_config = config_dir.join("config.json");
         if user_config.exists() {
             // Load a custom configuration file
-            xpg::Config::from_path(&user_config)?
+            xpg::Config::from_path(&user_config, cli.extended)?
         } else {
-            // Use the default configuration
-            xpg::CONFIG.clone()
+            // Load the default configuration
+            xpg::Config::from_str(xpg::CONFIG, cli.extended)?
         }
     } else {
-        // Use the default configuration
-        xpg::CONFIG.clone()
+        // Load the default configuration
+        xpg::Config::from_str(xpg::CONFIG, cli.extended)?
     };
+
+    if cli.dump_config {
+        print!("{}", cfg.dump()?);
+        std::process::exit(0);
+    } else if cli.list {
+        let patterns = if cli.patterns.is_empty() {
+            if cli.extended {
+                vec![String::from("{a.ext}")]
+            } else {
+                vec![String::from("{a}")]
+            }
+        } else {
+            cli.patterns.clone()
+        };
+        for sub in &patterns {
+            for word in &cfg.list(sub)? {
+                println!("{word}");
+            }
+        }
+        std::process::exit(0);
+    }
 
     // Convert --length, --minimum, --maximum option values into a range
     let length = if let Some(n) = cli.length {
@@ -123,66 +151,72 @@ fn main() -> Result<()> {
         0..=usize::MAX
     };
 
-    // Pad pattern to the minimum length with last character of pattern if it doesn't have any words
-    let min = length.start();
-    let len = cli.pattern.len();
-    let d = if len < *min { min - len } else { 0 };
-    let pattern = if d > 0
-        && !["keychain", "codename", "codename-series"].contains(&cli.pattern.as_str())
-        && !cli.pattern.contains(['w', 'W', 'T'])
-        && !xpg::WORD_KIND_SUBS
-            .iter()
-            .any(|(x, _)| cli.pattern.contains(x))
-    {
-        format!(
-            "{}{}",
-            cli.pattern,
-            format!("{}", cli.pattern.chars().last().unwrap()).repeat(d),
-        )
+    let patterns = if cli.patterns.is_empty() {
+        vec![String::from("wwww")]
     } else {
-        cli.pattern.clone()
+        cli.patterns.clone()
     };
 
-    // Adjust meaning of -c for codename-series
-    let (count, codenames) = if pattern == "codename-series" {
-        (1, if cli.count == 1 { 10 } else { cli.count })
-    } else {
-        (cli.count, 0)
-    };
+    for pattern in &patterns {
+        // Pad pattern to the minimum length with last character of pattern if it doesn't have any words
+        let min = length.start();
+        let len = pattern.len();
+        let d = if len < *min { min - len } else { 0 };
+        let pattern = if d > 0
+            && !["keychain", "codename", "codename-series"].contains(&pattern.as_str())
+            && !pattern.contains(['w', 'W', 'T'])
+            && !xpg::WORD_KIND_SUBS.keys().any(|x| pattern.contains(x))
+        {
+            format!(
+                "{}{}",
+                pattern,
+                format!("{}", pattern.chars().last().unwrap()).repeat(d),
+            )
+        } else {
+            pattern.clone()
+        };
 
-    // Generate password(s)
-    let mut i = 0;
-    loop {
-        let mut attempts = 0;
+        // Adjust meaning of -c for codename-series
+        let (count, codenames) = if pattern == "codename-series" {
+            (1, if cli.count == 1 { 10 } else { cli.count })
+        } else {
+            (cli.count, 0)
+        };
+
+        // Generate password(s)
+        let mut i = 0;
         loop {
-            let pw = match pattern.as_str() {
-                "keychain" => cfg.keychain(),
-                "codename" => cfg.codename(),
-                "codename-series" => cfg.codename_series(codenames),
-                _ => cfg.generate(&pattern),
-            };
+            let mut attempts = 0;
+            loop {
+                let pw = match pattern.as_str() {
+                    "keychain" => cfg.keychain(),
+                    "codename" => cfg.codename(),
+                    "codename-series" => cfg.codename_series(codenames),
+                    _ => cfg.generate(&pattern),
+                };
 
-            // Enforce length requirements
-            if length.contains(&pw.len()) {
-                println!("{pw}");
-                break;
-            } else {
-                attempts += 1;
-                if attempts >= cli.attempts {
-                    exit(1, "Maximum attempts exceeded!");
+                // Enforce length requirements
+                if length.contains(&pw.len()) {
+                    println!("{pw}");
+                    break;
+                } else {
+                    attempts += 1;
+                    if attempts >= cli.attempts {
+                        error("Maximum attempts exceeded!");
+                    }
                 }
             }
-        }
-        i += 1;
-        if count > 0 && i >= count {
-            break;
+            i += 1;
+            if count > 0 && i >= count {
+                break;
+            }
         }
     }
 
     Ok(())
 }
 
-fn exit(code: i32, msg: &str) {
-    eprintln!("ERROR: {msg}");
-    std::process::exit(code);
+fn error(msg: &str) {
+    eprintln!("Error: {msg}");
+    std::process::exit(1);
 }